middleware/requirements.txt
# Core
fastapi>=0.111.0
uvicorn>=0.29.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# LLM
google-generativeai>=0.5.0

# Caching
redis>=5.0.0

# Database
pymongo>=4.6.0

# Privacy & NER
spacy>=3.7.0
# After install run: python -m spacy download en_core_web_sm

# Encryption
cryptography>=42.0.0

# Auth
PyJWT>=2.8.0

# Scheduling
apscheduler>=3.10.0

# Google OAuth (for real Classroom integration)
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
anyio>=4.0.0
httpx>=0.27.0
